… opm/simulators/linalg/bda/Misc.cpp
… opm/simulators/linalg/bda/Preconditioner.hpp
… opm/simulators/linalg/bda/WellContributions.cpp
… opm/simulators/linalg/bda/cuda/cusparseSolverBackend.cu
… opm/simulators/linalg/bda/opencl/Preconditioner.cpp
… opm/simulators/linalg/bda/opencl/openclBILU0.cpp
… opm/simulators/linalg/bda/opencl/openclBILU0.hpp
… opm/simulators/linalg/bda/opencl/openclBISAI.cpp
… opm/simulators/linalg/bda/opencl/openclBISAI.hpp
… opm/simulators/linalg/bda/opencl/openclCPR.cpp
… opm/simulators/linalg/bda/opencl/openclCPR.hpp
… opm/simulators/linalg/bda/opencl/openclKernels.cpp
… opm/simulators/linalg/bda/opencl/openclPreconditioner.cpp
… opm/simulators/linalg/bda/opencl/openclPreconditioner.hpp
… opm/simulators/linalg/bda/opencl/openclSolverBackend.cpp
… opm/simulators/linalg/bda/opencl/openclSolverBackend.hpp
… opm/simulators/linalg/bda/rocm/hipKernels.cpp
… opm/simulators/linalg/bda/rocm/hipKernels.hpp
… opm/simulators/linalg/bda/rocm/rocalutionSolverBackend.cpp
… opm/simulators/linalg/bda/rocm/rocalutionSolverBackend.hpp
… opm/simulators/linalg/bda/rocm/rocsparseBILU0.cpp
… opm/simulators/linalg/bda/rocm/rocsparseBILU0.hpp
… opm/simulators/linalg/bda/rocm/rocsparseCPR.cpp
… opm/simulators/linalg/bda/rocm/rocsparseCPR.hpp
… opm/simulators/linalg/bda/rocm/rocsparseMatrix.cpp
… opm/simulators/linalg/bda/rocm/rocsparseMatrix.hpp
… opm/simulators/linalg/bda/rocm/rocsparsePreconditioner.cpp
… opm/simulators/linalg/bda/rocm/rocsparsePreconditioner.hpp
… opm/simulators/linalg/bda/rocm/rocsparseSolverBackend.cpp
… opm/simulators/linalg/bda/rocm/rocsparseSolverBackend.hpp