This PR adds support for applying CPR on AMD GPU using the rocsparse library. Similarly to the opencl implementation, the creation of AMG is done using the DUNE library.

Benchmark: NORNE
System:    CPU: AMD EPYC 7763 64-Core, 2.45GHz, GPU: AMD MI210

---

### Results ###

Units of measurement
* Total time [seconds]
* Assembly time [seconds]
* Linear solve time [seconds]
* Update time [seconds]
* Pre/post step [seconds]
* Overall Linearizations [count]
* Overall Newton Iterations [count]
* Overall Linear Iterations [count]

Case | Total Time | Assembly | Linear Solve | Update | Pre/Post | Linearisations | Newton | Linear Iterations
----- | ----------: | ----------: | -------------: | --------: | ---------: | --------------: | --------: | ---------------:
ilu0, jacobiBlocks=0, -coupled | 333 |74 | **97** | 76 | 82 | 1952 | 1582 | 22155
ilu0, jacobiBlocks=150, -coupled | 290 | 77 | **55** | 74 | 80 | 1951 | 1583 | 26336
cpr, jacobiBlocks=0, -coupled | 270 | 60 | **71** | 58 | 76 | 1571 | 1222 | 3273
cpr, jacobiBlocks=150, -coupled | 279 | 65 | **57** | 70 | 84 | 1560 | 1211 | 3557