@blattms 2024-06-27
Unfortunately there are compile errors on jenkins:
```
[236/478] Building CXX object CMakeFiles/opmsimulators.dir/opm/simulators/linalg/bda/CprCreation.cpp.o
FAILED: CMakeFiles/opmsimulators.dir/opm/simulators/linalg/bda/CprCreation.cpp.o 
/usr/lib/ccache/g++-10 -DBOOST_ALL_NO_LIB -DBOOST_ATOMIC_DYN_LINK -DBOOST_CHRONO_DYN_LINK -DBOOST_DATE_TIME_DYN_LINK -DBOOST_FILESYSTEM_DYN_LINK -DBOOST_LOG_DYN_LINK -DBOOST_LOG_SETUP_DYN_LINK -DBOOST_REGEX_DYN_LINK -DBOOST_SYSTEM_DYN_LINK -DBOOST_THREAD_DYN_LINK -DCL_HPP_MINIMUM_OPENCL_VERSION=120 -DCL_HPP_TARGET_OPENCL_VERSION=120 -DCL_TARGET_OPENCL_VERSION=120 -DENABLE_TBB=1 -DFMT_LOCALE -DFMT_SHARED -DHAVE_CONFIG_H=1 -DMETIS_API_VERSION=5 -DSCOTCH_METIS_VERSION=5 -DVEXCL_BACKEND_OPENCL -DVEXCL_HAVE_OPENCL_HPP -D_TBB_CPP0X -I/var/lib/jenkins/workspace/opm-simulators-PR-builder/mpi/build-opm-simulators -I/var/lib/jenkins/workspace/opm-simulators-PR-builder -I/usr/lib/x86_64-linux-gnu/openmpi/include/openmpi -I/usr/lib/x86_64-linux-gnu/openmpi/include -I/usr/include/parmetis -I/usr/include/superlu -I/usr/include/suitesparse -I/var/lib/jenkins/workspace/opm-simulators-PR-builder/mpi/install/include -I/usr/include/cjson -I/usr/include/trilinos -I/usr/include/scotch -isystem /var/lib/jenkins/amgcl/install/include -isystem /usr/include/metis -isystem /var/lib/jenkins/damaris/install/include -isystem /usr/include/hdf5/openmpi -pipe -Wall -Wextra -Wshadow -fopenmp -pthread -O3 -DNDEBUG -mtune=native -UNDEBUG -fPIC -fopenmp -std=c++17 -MD -MT CMakeFiles/opmsimulators.dir/opm/simulators/linalg/b[236/478] Building CXX object CMakeFiles/opmsimulators.dir/opm/simulators/linalg/bda/CprCreation.cpp.o
FAILED: CMakeFiles/opmsimulators.dir/opm/simulators/linalg/bda/CprCreation.cpp.o 
/usr/lib/ccache/g++-10 -DBOOST_ALL_NO_LIB -DBOOST_ATOMIC_DYN_LINK -DBOOST_CHRONO_DYN_LINK -DBOOST_DATE_TIME_DYN_LINK -DBOOST_FILESYSTEM_DYN_LINK -DBOOST_LOG_DYN_LINK -DBOOST_LOG_SETUP_DYN_LINK -DBOOST_REGEX_DYN_LINK -DBOOST_SYSTEM_DYN_LINK -DBOOST_THREAD_DYN_LINK -DCL_HPP_MINIMUM_OPENCL_VERSION=120 -DCL_HPP_TARGET_OPENCL_VERSION=120 -DCL_TARGET_OPENCL_VERSION=120 -DENABLE_TBB=1 -DFMT_LOCALE -DFMT_SHARED -DHAVE_CONFIG_H=1 -DMETIS_API_VERSION=5 -DSCOTCH_METIS_VERSION=5 -DVEXCL_BACKEND_OPENCL -DVEXCL_HAVE_OPENCL_HPP -D_TBB_CPP0X -I/var/lib/jenkins/workspace/opm-simulators-PR-builder/mpi/build-opm-simulators -I/var/lib/jenkins/workspace/opm-simulators-PR-builder -I/usr/lib/x86_64-linux-gnu/openmpi/include/openmpi -I/usr/lib/x86_64-linux-gnu/openmpi/include -I/usr/include/parmetis -I/usr/include/superlu -I/usr/include/suitesparse -I/var/lib/jenkins/workspace/opm-simulators-PR-builder/mpi/install/include -I/usr/include/cjson -I/usr/include/trilinos -I/usr/include/scotch -isystem /var/lib/jenkins/amgcl/install/include -isystem /usr/include/metis -isystem /var/lib/jenkins/damaris/install/include -isystem /usr/include/hdf5/openmpi -pipe -Wall -Wextra -Wshadow -fopenmp -pthread -O3 -DNDEBUG -mtune=native -UNDEBUG -fPIC -fopenmp -std=c++17 -MD -MT CMakeFiles/opmsimulators.dir/opm/simulators/linalg/bda/CprCreation.cpp.o -MF CMakeFiles/opmsimulators.dir/opm/simulators/linalg/bda/CprCreation.cpp.o.d -o CMakeFiles/opmsimulators.dir/opm/simulators/linalg/bda/CprCreation.cpp.o -c /var/lib/jenkins/workspace/opm-simulators-PR-builder/opm/simulators/linalg/bda/CprCreation.cpp
In file included from /var/lib/jenkins/workspace/opm-simulators-PR-builder/opm/simulators/linalg/bda/CprCreation.cpp:35:
/var/lib/jenkins/workspace/opm-simulators-PR-builder/opm/simulators/linalg/bda/Misc.hpp:4:10: fatal error: hip/hip_runtime_api.h: No such file or directory
    4 | #include <hip/hip_runtime_api.h>
      |          ^~~~~~~~~~~~~~~~~~~~~~~
compilation terminated.da/CprCreation.cpp.o -MF CMakeFiles/opmsimulators.dir/opm/simulators/linalg/bda/CprCreation.cpp.o.d -o CMakeFiles/opmsimulators.dir/opm/simulators/linalg/bda/CprCreation.cpp.o -c /var/lib/jenkins/workspace/opm-simulators-PR-builder/opm/simulators/linalg/bda/CprCreation.cpp
In file included from /var/lib/jenkins/workspace/opm-simulators-PR-builder/opm/simulators/linalg/bda/CprCreation.cpp:35:
/var/lib/jenkins/workspace/opm-simulators-PR-builder/opm/simulators/linalg/bda/Misc.hpp:4:10: fatal error: hip/hip_runtime_api.h: No such file or directory
    4 | #include <hip/hip_runtime_api.h>
      |          ^~~~~~~~~~~~~~~~~~~~~~~
compilation terminated.
```