@razvnane 2024-06-27
> This is surprisingly short. Was the implementation of CPR for rocSparse in a previous PR?

I am not sure what you mean. I just ported the opencl CPR implementation that Tong Dong did and I reused the AMG creation with dune and put it in a new class CprCreation. But everything related to CPR is in this PR.